+++ bedoner/lang/trf_mixin.py
@@ -1,14 +1,14 @@
 """The module trf_mixin defindes Language mixin for pytorch transformers."""
 from typing import Optional, Type
 
-import bedoner.lang.juman as juman
-from torch.optim.optimizer import Optimizer
-from bedoner.lang import torch_mixin
-from bedoner.torch_utils import OptimizerParameters
-from transformers import AdamW, WarmupLinearSchedule
 from spacy.tokens import Doc
 from spacy_transformers.language import TransformersLanguage
 from spacy_transformers.util import ATTRS
+from torch.optim.optimizer import Optimizer
+from transformers import AdamW, WarmupLinearSchedule
+
+from bedoner.lang import torch_mixin
+from bedoner.torch_utils import OptimizerParameters
 
 
 class TransformersLanguageMixin(torch_mixin.TorchLanguageMixin):
@@ -67,10 +67,6 @@ class TransformersLanguageMixin(torch_mixin.TorchLanguageMixin):
             optimizer, warmup_steps=warmup_steps, t_total=t_total
         )
         return torch_mixin.Optimizers(optimizer=optimizer, lr_scheduler=scheduler)
-
-
-class TransformersJuman(TransformersLanguageMixin, juman.Japanese):
-    """Juman language to manage pytorch-transformers components"""
 
 
 TransformersLanguageMixin.install_extensions()