+++ bedoner/models.py
@@ -2,18 +2,20 @@
 import os
 from pathlib import Path
 
-import bedoner.lang.juman as juman
-import bedoner.lang.mecab as mecab
-import bedoner.lang.knp as knp
 import mojimoji
-from bedoner.lang.trf_mixin import TransformersJuman
-from bedoner.pipelines.date_ner import DateRuler
+from spacy.language import Language
+from spacy.vocab import Vocab
+
+import bedoner.lang.juman as juman
+import bedoner.lang.knp as knp
+import bedoner.lang.mecab as mecab
+from bedoner.lang.trf_mixin import TransformersLanguageMixin
+from bedoner.pipelines.knp_ner import KnpEntityExtractor
 from bedoner.pipelines.person_ner import create_person_ruler
 from bedoner.pipelines.trf_model import BertModel
 from bedoner.pipelines.trf_ner import BertForNamedEntityRecognition
-from bedoner.pipelines.knp_ner import KnpEntityExtractor
 from bedoner.pipelines.wordpiecer import WordPiecer
-from spacy.vocab import Vocab
+from bedoner.utils import inject_mixin
 
 __dir__ = Path(__file__).parent
 
@@ -30,34 +32,31 @@ def juman_nlp() -> juman.Japanese:
     )
 
 
-def bert_wordpiecer() -> mecab.Japanese:
-    nlp = TransformersJuman(
-        Vocab(), meta={"tokenizer": {"preprocessor": han_to_zen_normalizer}}
-    )
-    w = WordPiecer.from_pretrained(Vocab(), bert_dir)
+def bert_wordpiecer(lang="juman", bert_dir=bert_dir) -> Language:
+    if lang == "juman":
+        cls = inject_mixin(TransformersLanguageMixin, juman.Japanese)
+        nlp = cls(Vocab(), meta={"tokenizer": {"preprocessor": han_to_zen_normalizer}})
+    elif lang == "mecab":
+        cls = inject_mixin(TransformersLanguageMixin, mecab.Japanese)
+        nlp = cls(Vocab())
+    else:
+        raise ValueError(f"Unsupported lang: {lang}")
+    w = WordPiecer.from_pretrained(nlp.vocab, bert_dir)
     nlp.add_pipe(w)
     return nlp
 
 
-def bert_model():
-    nlp = bert_wordpiecer()
+def bert_model(lang="juman", bert_dir=bert_dir):
+    nlp = bert_wordpiecer(lang)
     bert = BertModel.from_pretrained(Vocab(), bert_dir)
     nlp.add_pipe(bert)
     return nlp
 
 
-def bert_ner(**cfg):
-    nlp = bert_model()
+def bert_ner(lang="juman", bert_dir=bert_dir, **cfg):
+    nlp = bert_model(lang)
     ner = BertForNamedEntityRecognition.from_pretrained(Vocab(), bert_dir, **cfg)
     nlp.add_pipe(ner)
-    return nlp
-
-
-def date_ruler(name="date_ruler") -> mecab.Japanese:
-    nlp = mecab.Japanese(
-        meta={"name": name, "requirements": ["mecab-python3", "regex"]}
-    )
-    nlp.add_pipe(DateRuler())
     return nlp