+++ bedoner/pipelines/person_ner.py
@@ -1,6 +1,7 @@
-import bedoner.ner_labels.labels_ontonotes as L
 from spacy.language import Language
 from spacy.pipeline.entityruler import EntityRuler
+
+import bedoner.ner_labels.labels_ontonotes as L
 
 person_patterns = [
     {"label": L.PERSON, "pattern": [{"TAG": "名詞,固有名詞,人名,姓"}, {"TAG": "名詞,固有名詞,人名,名"}]},