+++ tests/conftest.py
@@ -1,9 +1,13 @@
-from .utils import check_juman, check_knp, check_mecab
+from pathlib import Path
+
 import pytest
+
 from bedoner.lang.juman import Japanese as Juman
 from bedoner.lang.knp import Japanese as KNP
 from bedoner.lang.mecab import Japanese as Mecab
 from bedoner.models import bert_wordpiecer
+
+from .utils import check_juman, check_knp, check_mecab
 
 
 @pytest.fixture(scope="session")
@@ -32,3 +36,8 @@ def knp_tokenizer():
 @pytest.fixture(scope="session")
 def bert_wordpiece_nlp():
     return bert_wordpiecer()
+
+
+@pytest.fixture(scope="session")
+def DATADIR():
+    return Path(__file__).parent / "data/"