+++ tests/pipelines/test_person.py
@@ -1,9 +1,10 @@
-import bedoner.ner_labels.labels_ontonotes as L
 import pytest
-from bedoner.lang.mecab import Japanese
-from bedoner.pipelines.person_ner import create_person_ruler
 from spacy.tokens.doc import Doc
 from spacy.tokens.span import Span
+
+import bedoner.ner_labels.labels_ontonotes as L
+from bedoner.lang.mecab import Japanese
+from bedoner.pipelines.person_ner import create_person_ruler
 
 TESTS = [("今日は高松隆と海に行った", "高松隆"), ("今日は田中と海に行った", "田中")]